Improved packages

+++ basic_settings/manifests/puppet.pp
@@ -27,14 +27,16 @@ class basic_settings::puppet(
     }
 
     /* Create drop in for puppet service */
-    basic_settings::systemd_drop_in { 'puppet_settings':
-        target_unit     => 'puppet.service',
-        unit            => {
-            'OnFailure' => 'notify-failed@%i.service'
-        },
-        service         => {
-            'Nice'          => 19,
-            'LimitNOFILE'   => 10000
+    if (defined(Package['systemd'])) {
+        basic_settings::systemd_drop_in { 'puppet_settings':
+            target_unit     => 'puppet.service',
+            unit            => {
+                'OnFailure' => 'notify-failed@%i.service'
+            },
+            service         => {
+                'Nice'          => 19,
+                'LimitNOFILE'   => 10000
+            }
         }
     }
 
@@ -57,42 +59,44 @@ class basic_settings::puppet(
             }
         }
 
-        /* Create drop in for puppet server service */
-        basic_settings::systemd_drop_in { 'puppetserver_settings':
-            target_unit     => "${server_package}.service",
-            unit            => {
-                'OnFailure' => 'notify-failed@%i.service'
-            },
-            service         => {
-                'Nice'          => '-8',
+        if (defined(Package['systemd'])) {
+            /* Create drop in for puppet server service */
+            basic_settings::systemd_drop_in { 'puppetserver_settings':
+                target_unit     => "${server_package}.service",
+                unit            => {
+                    'OnFailure' => 'notify-failed@%i.service'
+                },
+                service         => {
+                    'Nice'          => '-8',
+                }
             }
-        }
 
-        /* Create systemd puppet server clean reports service */
-        basic_settings::systemd_service { 'puppetserver-clean-reports':
-            description => 'Clean puppetserver reports service',
-            service     => {
-                'Type'      => 'oneshot',
-                'User'      => 'puppet',
-                'ExecStart' => "/usr/bin/find /var/lib/${server_dir}/reports -type f -name '*.yaml' -ctime +1 -delete",
-                'Nice'      => '19'
-            },
-        }
-
-        /* Create systemd puppet server clean reports timer */
-        basic_settings::systemd_timer { 'puppetserver-clean-reports':
-            description => 'Clean puppetserver reports timer',
-            timer       => {
-                'OnCalendar' => '*-*-* 10:00'
+            /* Create systemd puppet server clean reports service */
+            basic_settings::systemd_service { 'puppetserver-clean-reports':
+                description => 'Clean puppetserver reports service',
+                service     => {
+                    'Type'      => 'oneshot',
+                    'User'      => 'puppet',
+                    'ExecStart' => "/usr/bin/find /var/lib/${server_dir}/reports -type f -name '*.yaml' -ctime +1 -delete",
+                    'Nice'      => '19'
+                },
             }
-        }
 
-        /* Create drop in for puppet service */
-        basic_settings::systemd_drop_in { 'puppet_puppetserver_dependency':
-            target_unit     => 'puppet.service',
-            unit         => {
-                'After'     => "${server_package}.service",
-                'BindsTo'   => "${server_package}.service"
+            /* Create systemd puppet server clean reports timer */
+            basic_settings::systemd_timer { 'puppetserver-clean-reports':
+                description => 'Clean puppetserver reports timer',
+                timer       => {
+                    'OnCalendar' => '*-*-* 10:00'
+                }
+            }
+
+            /* Create drop in for puppet service */
+            basic_settings::systemd_drop_in { 'puppet_puppetserver_dependency':
+                target_unit     => 'puppet.service',
+                unit         => {
+                    'After'     => "${server_package}.service",
+                    'BindsTo'   => "${server_package}.service"
+                }
             }
         }
     }
